[POC-15] add read matches function
