+++ src/generative_dating_agents/database/chroma.py
@@ -14,7 +14,6 @@ from chromadb.api.models.Collection import (
 from chromadb.config import Settings
 from chromadb.utils import embedding_functions
 
-from ..data.models import UserProfile
 from ..utils.constants import (
     CHROMA_DISTANCE,
     CHROMA_DISTANCE_KEY,
@@ -22,6 +21,7 @@ from ..utils.constants import (
     CHROMA_USER_PROFILES_COLLECTION_NAME,
 )
 from ..utils.io import read_all_user_profiles
+from ..utils.types import UserProfile
 
 
 class ChromaVectorDatabaseClient:
